add: add search function filter and move to custome hook
